requirements.txt
SQLAlchemy
uvicorn[standard]
fastapi
requests
python-decouple
psycopg2-binary